+++ account/views.py
@@ -4,8 +4,46 @@ from rest_framework.viewsets import ModelViewSet
 from account.serializers import AccountSerializer
 
 from database.models import UserDatas
+from django.contrib.auth.backends import ModelBackend
+from django.contrib.auth import get_user_model
+from django.db.models import Q
+
+UserModel = get_user_model()
 
 class AccountViewSet(ModelViewSet):
     queryset = UserDatas.objects.all()
     serializer_class = AccountSerializer
+
+
+
+
+class EmailBackend(ModelBackend):
+    def authenticate(self, request, username=None, password=None, **kwargs):
+        try:
+            user = UserModel.objects.get(Q(usernameAccount=username) | Q(emailAccount=username))
+        except UserModel.DoesNotExist:
+            UserModel().set_password(password)
+            return
+        except UserModel.MultipleObjectsReturned:
+            user = UserModel.objects.filter(Q(usernameAccount=username) | Q(emailAccount=username))
+
+        if user.check_password(password) and self.user_can_authenticate(user):
+            return user
+
+    # def authenticate(self, request, username=None, password=None, **kwargs):
+    #     if username is None:
+    #         username = kwargs.get(UserModel.USERNAME_FIELD)
+    #     try:
+    #         # user = UserModel._default_manager.get_by_natural_key(username)
+    #         # You can customise what the given username is checked against, here I compare to both 
+    #         # username and email fields of the User model
+    #         user = UserModel.objects.get(Q(usernameAccount=username) | Q(emailAccount=username))
+    #     except UserModel.DoesNotExist:
+    #         # Run the default password hasher once to reduce the timing
+    #         # difference between an existing and a nonexistent user (#20760).
+    #         UserModel().set_password(password)
+    #     else:
+    #         if user.check_password(password) and self.user_can_authenticate(user):
+    #             return user
+    #     return super().authenticate(request, username, password, **kwargs)
